Add files utils, plugin and serialization for v6.1
