+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/overlays/OverlaySettingsSerialization.java
@@ -1,9 +1,9 @@
-package com.microblink.flutter.overlays;
+package com.microblink.blinkid.flutter.overlays;
 
 import android.content.Context;
 
-import com.microblink.entities.recognizers.RecognizerBundle;
-import com.microblink.uisettings.UISettings;
+import com.microblink.blinkid.entities.recognizers.RecognizerBundle;
+import com.microblink.blinkid.uisettings.UISettings;
 
 import org.json.JSONObject;
