+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/overlays/serialization/DocumentOverlaySettingsSerialization.java
@@ -1,11 +1,11 @@
-package com.microblink.flutter.overlays.serialization;
+package com.microblink.blinkid.flutter.overlays.serialization;
 
 import android.content.Context;
 
-import com.microblink.entities.recognizers.RecognizerBundle;;
-import com.microblink.uisettings.DocumentUISettings;
-import com.microblink.uisettings.UISettings;
-import com.microblink.flutter.overlays.OverlaySettingsSerialization;
+import com.microblink.blinkid.entities.recognizers.RecognizerBundle;;
+import com.microblink.blinkid.uisettings.DocumentUISettings;
+import com.microblink.blinkid.uisettings.UISettings;
+import com.microblink.blinkid.flutter.overlays.OverlaySettingsSerialization;
 
 import org.json.JSONObject;
