+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/recognizers/serialization/BlinkIdMultiSideRecognizerSerialization.java
@@ -1,22 +1,22 @@
-package com.microblink.flutter.recognizers.serialization;
+package com.microblink.blinkid.flutter.recognizers.serialization;
 
-import com.microblink.entities.recognizers.Recognizer;
-import com.microblink.flutter.recognizers.RecognizerSerialization;
-import com.microblink.flutter.SerializationUtils;
+import com.microblink.blinkid.entities.recognizers.Recognizer;
+import com.microblink.blinkid.flutter.recognizers.RecognizerSerialization;
+import com.microblink.blinkid.flutter.SerializationUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public final class BlinkIdCombinedRecognizerSerialization implements RecognizerSerialization {
+public final class BlinkIdMultiSideRecognizerSerialization implements RecognizerSerialization {
 
     @Override
     public Recognizer<?> createRecognizer(JSONObject jsonObject) {
-        com.microblink.entities.recognizers.blinkid.generic.BlinkIdCombinedRecognizer recognizer = new com.microblink.entities.recognizers.blinkid.generic.BlinkIdCombinedRecognizer();
+        com.microblink.blinkid.entities.recognizers.blinkid.generic.BlinkIdMultiSideRecognizer recognizer = new com.microblink.blinkid.entities.recognizers.blinkid.generic.BlinkIdMultiSideRecognizer();
         recognizer.setAllowBlurFilter(jsonObject.optBoolean("allowBlurFilter", true));
         recognizer.setAllowUncertainFrontSideScan(jsonObject.optBoolean("allowUncertainFrontSideScan", false));
         recognizer.setAllowUnparsedMrzResults(jsonObject.optBoolean("allowUnparsedMrzResults", false));
         recognizer.setAllowUnverifiedMrzResults(jsonObject.optBoolean("allowUnverifiedMrzResults", true));
-        recognizer.setAnonymizationMode(com.microblink.entities.recognizers.blinkid.generic.AnonymizationMode.values()[jsonObject.optInt("anonymizationMode", 4) - 1]);
+        recognizer.setAnonymizationMode(com.microblink.blinkid.entities.recognizers.blinkid.generic.AnonymizationMode.values()[jsonObject.optInt("anonymizationMode", 4) - 1]);
         recognizer.setFaceImageDpi(jsonObject.optInt("faceImageDpi", 250));
         recognizer.setFullDocumentImageDpi(jsonObject.optInt("fullDocumentImageDpi", 250));
         recognizer.setFullDocumentImageExtensionFactors(SerializationUtils.deserializeExtensionFactors(jsonObject.optJSONObject("fullDocumentImageExtensionFactors")));
@@ -36,15 +36,16 @@ public final class BlinkIdCombinedRecognizerSerialization implements RecognizerS
 
     @Override
     public JSONObject serializeResult(Recognizer<?> recognizer) {
-        com.microblink.entities.recognizers.blinkid.generic.BlinkIdCombinedRecognizer.Result result = ((com.microblink.entities.recognizers.blinkid.generic.BlinkIdCombinedRecognizer)recognizer).getResult();
+        com.microblink.blinkid.entities.recognizers.blinkid.generic.BlinkIdMultiSideRecognizer.Result result = ((com.microblink.blinkid.entities.recognizers.blinkid.generic.BlinkIdMultiSideRecognizer)recognizer).getResult();
         JSONObject jsonResult = new JSONObject();
         try {
             SerializationUtils.addCommonRecognizerResultData(jsonResult, result);
-            jsonResult.put("additionalAddressInformation", result.getAdditionalAddressInformation());
-            jsonResult.put("additionalNameInformation", result.getAdditionalNameInformation());
-            jsonResult.put("additionalOptionalAddressInformation", result.getAdditionalOptionalAddressInformation());
-            jsonResult.put("address", result.getAddress());
+            jsonResult.put("additionalAddressInformation", BlinkIDSerializationUtils.serializeStringResult(result.getAdditionalAddressInformation()));
+            jsonResult.put("additionalNameInformation", BlinkIDSerializationUtils.serializeStringResult(result.getAdditionalNameInformation()));
+            jsonResult.put("additionalOptionalAddressInformation", BlinkIDSerializationUtils.serializeStringResult(result.getAdditionalOptionalAddressInformation()));
+            jsonResult.put("address", BlinkIDSerializationUtils.serializeStringResult(result.getAddress()));
             jsonResult.put("age", result.getAge());
+            jsonResult.put("backAdditionalProcessingInfo", BlinkIDSerializationUtils.serializeAdditionalProcessingInfo(result.getBackAdditionalProcessingInfo()));
             jsonResult.put("backCameraFrame", SerializationUtils.encodeImageBase64(result.getBackCameraFrame()));
             jsonResult.put("backImageAnalysisResult", BlinkIDSerializationUtils.serializeImageAnalysisResult(result.getBackImageAnalysisResult()));
             jsonResult.put("backProcessingStatus", SerializationUtils.serializeEnum(result.getBackProcessingStatus()));
@@ -52,46 +53,47 @@ public final class BlinkIdCombinedRecognizerSerialization implements RecognizerS
             jsonResult.put("barcodeCameraFrame", SerializationUtils.encodeImageBase64(result.getBarcodeCameraFrame()));
             jsonResult.put("barcodeResult", BlinkIDSerializationUtils.serializeBarcodeResult(result.getBarcodeResult()));
             jsonResult.put("classInfo", BlinkIDSerializationUtils.serializeClassInfo(result.getClassInfo()));
-            jsonResult.put("dataMatchDetailedInfo", BlinkIDSerializationUtils.serializeDataMatchDetailedInfo(result.getDataMatchDetailedInfo()));
-            jsonResult.put("dateOfBirth", SerializationUtils.serializeDate(result.getDateOfBirth()));
-            jsonResult.put("dateOfExpiry", SerializationUtils.serializeDate(result.getDateOfExpiry()));
+            jsonResult.put("dataMatch", BlinkIDSerializationUtils.serializeDataMatchResult(result.getDataMatch()));
+            jsonResult.put("dateOfBirth", BlinkIDSerializationUtils.serializeDateResult(result.getDateOfBirth()));
+            jsonResult.put("dateOfExpiry", BlinkIDSerializationUtils.serializeDateResult(result.getDateOfExpiry()));
             jsonResult.put("dateOfExpiryPermanent", result.isDateOfExpiryPermanent());
-            jsonResult.put("dateOfIssue", SerializationUtils.serializeDate(result.getDateOfIssue()));
-            jsonResult.put("documentAdditionalNumber", result.getDocumentAdditionalNumber());
-            jsonResult.put("documentDataMatch", SerializationUtils.serializeEnum(result.getDocumentDataMatch()));
-            jsonResult.put("documentNumber", result.getDocumentNumber());
-            jsonResult.put("documentOptionalAdditionalNumber", result.getDocumentOptionalAdditionalNumber());
+            jsonResult.put("dateOfIssue", BlinkIDSerializationUtils.serializeDateResult(result.getDateOfIssue()));
+            jsonResult.put("documentAdditionalNumber", BlinkIDSerializationUtils.serializeStringResult(result.getDocumentAdditionalNumber()));
+            jsonResult.put("documentNumber", BlinkIDSerializationUtils.serializeStringResult(result.getDocumentNumber()));
+            jsonResult.put("documentOptionalAdditionalNumber", BlinkIDSerializationUtils.serializeStringResult(result.getDocumentOptionalAdditionalNumber()));
             jsonResult.put("driverLicenseDetailedInfo", BlinkIDSerializationUtils.serializeDriverLicenseDetailedInfo(result.getDriverLicenseDetailedInfo()));
-            jsonResult.put("employer", result.getEmployer());
+            jsonResult.put("employer", BlinkIDSerializationUtils.serializeStringResult(result.getEmployer()));
             jsonResult.put("expired", result.isExpired());
             jsonResult.put("faceImage", SerializationUtils.encodeImageBase64(result.getFaceImage()));
-            jsonResult.put("fathersName", result.getFathersName());
-            jsonResult.put("firstName", result.getFirstName());
+            jsonResult.put("fathersName", BlinkIDSerializationUtils.serializeStringResult(result.getFathersName()));
+            jsonResult.put("firstName", BlinkIDSerializationUtils.serializeStringResult(result.getFirstName()));
+            jsonResult.put("frontAdditionalProcessingInfo", BlinkIDSerializationUtils.serializeAdditionalProcessingInfo(result.getFrontAdditionalProcessingInfo()));
             jsonResult.put("frontCameraFrame", SerializationUtils.encodeImageBase64(result.getFrontCameraFrame()));
             jsonResult.put("frontImageAnalysisResult", BlinkIDSerializationUtils.serializeImageAnalysisResult(result.getFrontImageAnalysisResult()));
             jsonResult.put("frontProcessingStatus", SerializationUtils.serializeEnum(result.getFrontProcessingStatus()));
             jsonResult.put("frontVizResult", BlinkIDSerializationUtils.serializeVizResult(result.getFrontVizResult()));
             jsonResult.put("fullDocumentBackImage", SerializationUtils.encodeImageBase64(result.getFullDocumentBackImage()));
             jsonResult.put("fullDocumentFrontImage", SerializationUtils.encodeImageBase64(result.getFullDocumentFrontImage()));
-            jsonResult.put("fullName", result.getFullName());
-            jsonResult.put("issuingAuthority", result.getIssuingAuthority());
-            jsonResult.put("lastName", result.getLastName());
-            jsonResult.put("localizedName", result.getLocalizedName());
-            jsonResult.put("maritalStatus", result.getMaritalStatus());
-            jsonResult.put("mothersName", result.getMothersName());
+            jsonResult.put("fullName", BlinkIDSerializationUtils.serializeStringResult(result.getFullName()));
+            jsonResult.put("issuingAuthority", BlinkIDSerializationUtils.serializeStringResult(result.getIssuingAuthority()));
+            jsonResult.put("lastName", BlinkIDSerializationUtils.serializeStringResult(result.getLastName()));
+            jsonResult.put("localizedName", BlinkIDSerializationUtils.serializeStringResult(result.getLocalizedName()));
+            jsonResult.put("maritalStatus", BlinkIDSerializationUtils.serializeStringResult(result.getMaritalStatus()));
+            jsonResult.put("mothersName", BlinkIDSerializationUtils.serializeStringResult(result.getMothersName()));
             jsonResult.put("mrzResult", BlinkIDSerializationUtils.serializeMrzResult(result.getMrzResult()));
-            jsonResult.put("nationality", result.getNationality());
-            jsonResult.put("personalIdNumber", result.getPersonalIdNumber());
-            jsonResult.put("placeOfBirth", result.getPlaceOfBirth());
+            jsonResult.put("nationality", BlinkIDSerializationUtils.serializeStringResult(result.getNationality()));
+            jsonResult.put("personalIdNumber", BlinkIDSerializationUtils.serializeStringResult(result.getPersonalIdNumber()));
+            jsonResult.put("placeOfBirth", BlinkIDSerializationUtils.serializeStringResult(result.getPlaceOfBirth()));
             jsonResult.put("processingStatus", SerializationUtils.serializeEnum(result.getProcessingStatus()));
-            jsonResult.put("profession", result.getProfession());
-            jsonResult.put("race", result.getRace());
+            jsonResult.put("profession", BlinkIDSerializationUtils.serializeStringResult(result.getProfession()));
+            jsonResult.put("race", BlinkIDSerializationUtils.serializeStringResult(result.getRace()));
             jsonResult.put("recognitionMode", SerializationUtils.serializeEnum(result.getRecognitionMode()));
-            jsonResult.put("religion", result.getReligion());
-            jsonResult.put("residentialStatus", result.getResidentialStatus());
+            jsonResult.put("religion", BlinkIDSerializationUtils.serializeStringResult(result.getReligion()));
+            jsonResult.put("residentialStatus", BlinkIDSerializationUtils.serializeStringResult(result.getResidentialStatus()));
             jsonResult.put("scanningFirstSideDone", result.isScanningFirstSideDone());
-            jsonResult.put("sex", result.getSex());
+            jsonResult.put("sex", BlinkIDSerializationUtils.serializeStringResult(result.getSex()));
             jsonResult.put("signatureImage", SerializationUtils.encodeImageBase64(result.getSignatureImage()));
+            jsonResult.put("version", (int)result.getVersion());
         } catch (JSONException e) {
             // see https://developer.android.com/reference/org/json/JSONException
             throw new RuntimeException(e);
@@ -101,11 +103,11 @@ public final class BlinkIdCombinedRecognizerSerialization implements RecognizerS
 
     @Override
     public String getJsonName() {
-        return "BlinkIdCombinedRecognizer";
+        return "BlinkIdMultiSideRecognizer";
     }
 
     @Override
     public Class<?> getRecognizerClass() {
-        return com.microblink.entities.recognizers.blinkid.generic.BlinkIdCombinedRecognizer.class;
+        return com.microblink.blinkid.entities.recognizers.blinkid.generic.BlinkIdMultiSideRecognizer.class;
     }
 }
