+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/recognizers/serialization/MrtdCombinedRecognizerSerialization.java
@@ -1,8 +1,8 @@
-package com.microblink.flutter.recognizers.serialization;
+package com.microblink.blinkid.flutter.recognizers.serialization;
 
-import com.microblink.entities.recognizers.Recognizer;
-import com.microblink.flutter.recognizers.RecognizerSerialization;
-import com.microblink.flutter.SerializationUtils;
+import com.microblink.blinkid.entities.recognizers.Recognizer;
+import com.microblink.blinkid.flutter.recognizers.RecognizerSerialization;
+import com.microblink.blinkid.flutter.SerializationUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -11,11 +11,11 @@ public final class MrtdCombinedRecognizerSerialization implements RecognizerSeri
 
     @Override
     public Recognizer<?> createRecognizer(JSONObject jsonObject) {
-        com.microblink.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer recognizer = new com.microblink.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer();
+        com.microblink.blinkid.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer recognizer = new com.microblink.blinkid.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer();
         recognizer.setAllowSpecialCharacters(jsonObject.optBoolean("allowSpecialCharacters", false));
         recognizer.setAllowUnparsedResults(jsonObject.optBoolean("allowUnparsedResults", false));
         recognizer.setAllowUnverifiedResults(jsonObject.optBoolean("allowUnverifiedResults", false));
-        recognizer.setDetectorType(com.microblink.entities.recognizers.blinkid.documentface.DocumentFaceDetectorType.values()[jsonObject.optInt("detectorType", 1) - 1]);
+        recognizer.setDetectorType(com.microblink.blinkid.entities.recognizers.blinkid.documentface.DocumentFaceDetectorType.values()[jsonObject.optInt("detectorType", 1) - 1]);
         recognizer.setFaceImageDpi(jsonObject.optInt("faceImageDpi", 250));
         recognizer.setFullDocumentImageDpi(jsonObject.optInt("fullDocumentImageDpi", 250));
         recognizer.setFullDocumentImageExtensionFactors(SerializationUtils.deserializeExtensionFactors(jsonObject.optJSONObject("fullDocumentImageExtensionFactors")));
@@ -27,7 +27,7 @@ public final class MrtdCombinedRecognizerSerialization implements RecognizerSeri
 
     @Override
     public JSONObject serializeResult(Recognizer<?> recognizer) {
-        com.microblink.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer.Result result = ((com.microblink.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer)recognizer).getResult();
+        com.microblink.blinkid.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer.Result result = ((com.microblink.blinkid.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer)recognizer).getResult();
         JSONObject jsonResult = new JSONObject();
         try {
             SerializationUtils.addCommonRecognizerResultData(jsonResult, result);
@@ -51,6 +51,6 @@ public final class MrtdCombinedRecognizerSerialization implements RecognizerSeri
 
     @Override
     public Class<?> getRecognizerClass() {
-        return com.microblink.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer.class;
+        return com.microblink.blinkid.entities.recognizers.blinkid.mrtd.MrtdCombinedRecognizer.class;
     }
 }
