+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/recognizers/serialization/UsdlRecognizerSerialization.java
@@ -1,41 +1,32 @@
-package com.microblink.flutter.recognizers.serialization;
+package com.microblink.blinkid.flutter.recognizers.serialization;
 
-import com.microblink.entities.recognizers.Recognizer;
-import com.microblink.entities.recognizers.blinkbarcode.usdl.UsdlKeys;
-import com.microblink.entities.recognizers.blinkid.usdl.UsdlCombinedRecognizer;
-import com.microblink.flutter.recognizers.RecognizerSerialization;
+import com.microblink.blinkid.entities.recognizers.Recognizer;
+import com.microblink.blinkid.entities.recognizers.blinkbarcode.usdl.UsdlKeys;
+import com.microblink.blinkid.entities.recognizers.blinkbarcode.usdl.UsdlRecognizer;
+import com.microblink.blinkid.flutter.recognizers.RecognizerSerialization;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import com.microblink.flutter.SerializationUtils;
+import com.microblink.blinkid.flutter.SerializationUtils;
 
-public final class UsdlCombinedRecognizerSerialization implements RecognizerSerialization {
+public final class UsdlRecognizerSerialization implements RecognizerSerialization {
 
     @Override
     public Recognizer<?> createRecognizer(JSONObject jsonRecognizer) {
-        UsdlCombinedRecognizer recognizer = new UsdlCombinedRecognizer();
-        recognizer.setFaceImageDpi(jsonRecognizer.optInt("faceImageDpi", 250));
-        recognizer.setFullDocumentImageDpi(jsonRecognizer.optInt("fullDocumentImageDpi", 250));
-        recognizer.setFullDocumentImageExtensionFactors(SerializationUtils.deserializeExtensionFactors(jsonRecognizer.optJSONObject("fullDocumentImageExtensionFactors")));
-        recognizer.setNumStableDetectionsThreshold(jsonRecognizer.optInt("numStableDetectionsThreshold", 6));
-        recognizer.setReturnFaceImage(jsonRecognizer.optBoolean("returnFaceImage", false));
-        recognizer.setReturnFullDocumentImage(jsonRecognizer.optBoolean("returnFullDocumentImage", false));
+        com.microblink.blinkid.entities.recognizers.blinkbarcode.usdl.UsdlRecognizer recognizer = new com.microblink.blinkid.entities.recognizers.blinkbarcode.usdl.UsdlRecognizer();
+        recognizer.setNullQuietZoneAllowed(jsonRecognizer.optBoolean("nullQuietZoneAllowed", true));
+        recognizer.setUncertainDecoding(jsonRecognizer.optBoolean("uncertainDecoding", true));
         return recognizer;
     }
 
     @Override
     public JSONObject serializeResult(Recognizer<?> recognizer) {
-        UsdlCombinedRecognizer.Result result = ((UsdlCombinedRecognizer)recognizer).getResult();
+        UsdlRecognizer.Result result = ((UsdlRecognizer)recognizer).getResult();
         JSONObject jsonResult = new JSONObject();
         try {
             SerializationUtils.addCommonRecognizerResultData(jsonResult, result);
-            jsonResult.put("documentDataMatch", SerializationUtils.serializeEnum(result.getDocumentDataMatch()));
-            jsonResult.put("faceImage", SerializationUtils.encodeImageBase64(result.getFaceImage()));
-            jsonResult.put("fullDocumentImage", SerializationUtils.encodeImageBase64(result.getFullDocumentImage()));
-            jsonResult.put("scanningFirstSideDone", result.isScanningFirstSideDone());
-
             jsonResult.put("optionalElements", SerializationUtils.serializeStringArray(result.getOptionalElements()));
             jsonResult.put("rawData", SerializationUtils.encodeByteArrayToBase64(result.getRawData()));
             jsonResult.put("rawStringData", result.getRawStringData());
@@ -63,7 +54,7 @@ public final class UsdlCombinedRecognizerSerialization implements RecognizerSeri
         return jsonResult;
     }
 
-    private JSONArray serializeFields(UsdlCombinedRecognizer.Result result) {
+    private JSONArray serializeFields(UsdlRecognizer.Result result) {
         JSONArray fieldsArr = new JSONArray();
         for (int i = 0; i < UsdlKeys.values().length; ++i) {
             fieldsArr.put(result.getField(UsdlKeys.values()[i]));
@@ -73,11 +64,11 @@ public final class UsdlCombinedRecognizerSerialization implements RecognizerSeri
 
     @Override
     public String getJsonName() {
-        return "UsdlCombinedRecognizer";
+        return "UsdlRecognizer";
     }
 
     @Override
     public Class<?> getRecognizerClass() {
-        return com.microblink.entities.recognizers.blinkid.usdl.UsdlCombinedRecognizer.class;
+        return com.microblink.blinkid.entities.recognizers.blinkbarcode.usdl.UsdlRecognizer.class;
     }
 }
