+++ BlinkID/android/src/main/java/com/microblink/blinkid/flutter/recognizers/serialization/VisaRecognizerSerialization.java
@@ -1,33 +1,33 @@
-package com.microblink.flutter.recognizers.serialization;
+package com.microblink.blinkid.flutter.recognizers.serialization;
 
-import com.microblink.entities.recognizers.Recognizer;
-import com.microblink.flutter.recognizers.RecognizerSerialization;
-import com.microblink.flutter.SerializationUtils;
+import com.microblink.blinkid.entities.recognizers.Recognizer;
+import com.microblink.blinkid.flutter.recognizers.RecognizerSerialization;
+import com.microblink.blinkid.flutter.SerializationUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public final class MrtdRecognizerSerialization implements RecognizerSerialization {
+public final class VisaRecognizerSerialization implements RecognizerSerialization {
 
     @Override
     public Recognizer<?> createRecognizer(JSONObject jsonObject) {
-        com.microblink.entities.recognizers.blinkid.mrtd.MrtdRecognizer recognizer = new com.microblink.entities.recognizers.blinkid.mrtd.MrtdRecognizer();
-        recognizer.setAllowSpecialCharacters(jsonObject.optBoolean("allowSpecialCharacters", false));
-        recognizer.setAllowUnparsedResults(jsonObject.optBoolean("allowUnparsedResults", false));
-        recognizer.setAllowUnverifiedResults(jsonObject.optBoolean("allowUnverifiedResults", false));
+        com.microblink.blinkid.entities.recognizers.blinkid.visa.VisaRecognizer recognizer = new com.microblink.blinkid.entities.recognizers.blinkid.visa.VisaRecognizer();
         recognizer.setDetectGlare(jsonObject.optBoolean("detectGlare", true));
+        recognizer.setFaceImageDpi(jsonObject.optInt("faceImageDpi", 250));
         recognizer.setFullDocumentImageDpi(jsonObject.optInt("fullDocumentImageDpi", 250));
         recognizer.setFullDocumentImageExtensionFactors(SerializationUtils.deserializeExtensionFactors(jsonObject.optJSONObject("fullDocumentImageExtensionFactors")));
+        recognizer.setReturnFaceImage(jsonObject.optBoolean("returnFaceImage", false));
         recognizer.setReturnFullDocumentImage(jsonObject.optBoolean("returnFullDocumentImage", false));
         return recognizer;
     }
 
     @Override
     public JSONObject serializeResult(Recognizer<?> recognizer) {
-        com.microblink.entities.recognizers.blinkid.mrtd.MrtdRecognizer.Result result = ((com.microblink.entities.recognizers.blinkid.mrtd.MrtdRecognizer)recognizer).getResult();
+        com.microblink.blinkid.entities.recognizers.blinkid.visa.VisaRecognizer.Result result = ((com.microblink.blinkid.entities.recognizers.blinkid.visa.VisaRecognizer)recognizer).getResult();
         JSONObject jsonResult = new JSONObject();
         try {
             SerializationUtils.addCommonRecognizerResultData(jsonResult, result);
+            jsonResult.put("faceImage", SerializationUtils.encodeImageBase64(result.getFaceImage()));
             jsonResult.put("fullDocumentImage", SerializationUtils.encodeImageBase64(result.getFullDocumentImage()));
             jsonResult.put("mrzResult", BlinkIDSerializationUtils.serializeMrzResult(result.getMrzResult()));
         } catch (JSONException e) {
@@ -39,11 +39,11 @@ public final class MrtdRecognizerSerialization implements RecognizerSerializatio
 
     @Override
     public String getJsonName() {
-        return "MrtdRecognizer";
+        return "VisaRecognizer";
     }
 
     @Override
     public Class<?> getRecognizerClass() {
-        return com.microblink.entities.recognizers.blinkid.mrtd.MrtdRecognizer.class;
+        return com.microblink.blinkid.entities.recognizers.blinkid.visa.VisaRecognizer.class;
     }
 }
